Install a conversion constant or offset between user-defined units.

Tells the units package how to convert between units that have a linear relationship, i.e. can be related on the form \(y = \alpha x\) (constant) or \(y = \alpha + x\) (offset).

install_conversion_constant(from, to, const)

install_conversion_offset(from, to, const)

Arguments

from
String for the symbol of the unit being converted from.

to
String for the symbol of the unit being converted to. One of from and to must be an existing unit name.

const
The constant \(\alpha\) in the conversion.

Details

This function handles the very common case where units are related through a linear function, that is, you can convert from one to the other as \(y = \alpha x\). Using this function, you specify that you can go from values of type from to values of type to by multiplying by a constant, or adding a constant.

See also

Examples

# one orange is worth two apples
install_symbolic_unit("orange")
install_conversion_constant("orange", "apple", 2) # apple = 2 * orange
apples <- 2 * as_units("apple")
oranges <- 1 * as_units("orange")
apples + oranges

#> 4 [apple]

oranges + apples

#> 2 [orange]

install_conversion_offset("meter", "newmeter", 1)
m = set_units(1:3, meter)
n = set_units(1:3, newmeter)
m + n

#> Units: [m]
#> [1] 3 5 7

n + m

#> Units: [newmeter]
#> [1] 1 3 5