create axis label with appropriate labels

create axis label with appropriate labels

plot unit objects

make_unit_label(lab, u, sep = units_options("sep"),
  group = units_options("group"), parse = units_options("parse"))

# S3 method for units
plot(x, y, xlab = NULL, ylab = NULL, ...)

Arguments

lab
length one character; name of the variable to plot

u
vector of class units

sep
length two character vector, defaulting to c("~","~"), with the white space between unit name and unit symbols, and between subsequent symbols.

group
length two character vector with grouping symbols, e.g. c("(",")") for parenthesis, or c("","") for no group symbols

parse
logical; indicates whether a parseable expression should be returned (typically needed for super scripts), or a simple character string without special formatting.

x
object of class units, to plot along the x axis, or, if y is missing, along the y axis

y
object to plot along the y axis, or missing

xlab
character; x axis label

ylab
character; y axis label

...
other parameters, passed on to plot.default

Details

units_options can be used to set and change the defaults for sep, group and doParse.

Examples

oldpar = par(mar = par("mar") + c(0, .3, 0, 0))
displacement = mtcars$disp * ud_units[["in"]]^3
# an example that would break if parse were (default) TRUE, since 'in' is a reserved word:
units_options(parse=FALSE)
make_unit_label("displacement", displacement)

#> Error in UseMethod("units"): no applicable method for 'units' applied to an object of class "c('double', 'numeric')"

units_options(parse=TRUE)
units(displacement) = with(ud_units, cm^3)

#> Error in cm^3: non-numeric argument to binary operator

weight = mtcars$wt * 1000 * with(ud_units, lb)

#> Error in eval(substitute(expr), data, enclos = parent.frame()): object 'lb' not found

units(weight) = with(ud_units, kg)

#> Error in eval(substitute(expr), data, enclos = parent.frame()): object 'kg' not found

plot(weight, displacement)

#> Error in plot(weight, displacement): object 'weight' not found

units_options(group = c("(", ")") )  # parenthesis instead of square brackets
plot(weight, displacement)

#> Error in plot(weight, displacement): object 'weight' not found

units_options(sep = c("~~~", "~"), group = c("", ""))  # no brackets; extra space
plot(weight, displacement)

#> Error in plot(weight, displacement): object 'weight' not found

units_options(sep = c("~", "~~"), group = c("[", "]"))
gallon = as_units("gallon")
consumption = mtcars$mpg * with(ud_units, mi/gallon)

#> Error in eval(substitute(expr), data, enclos = parent.frame()): object 'mi' not found

units(consumption) = with(ud_units, km/l)

#> Error in eval(substitute(expr), data, enclos = parent.frame()): object 'km' not found

plot(displacement, consumption) # division in consumption

#> Error in xy.coords(x, y, xlabel, ylabel, log): object 'consumption' not found

units_options(negative_power = TRUE) # division becomes ^-1
plot(displacement, consumption)

#> Error in xy.coords(x, y, xlabel, ylabel, log): object 'consumption' not found

plot(1/displacement, 1/consumption)

#> Error in xy.coords(x, y, xlabel, ylabel, log): object 'consumption' not found

par(oldpar)