set one or more units global options

set units global options, mostly related how units are printed and plotted

units_options(..., sep, group, negative_power, parse)

Arguments

...

ignored

sep

character length two; default c("~", "~"); space separator between variable and units, and space separator between two different units

group

character length two; start and end group, may be two empty strings, a parenthesis pair, or square brackets.

negative_power

logical, default FALSE; should denominators have negative power, or follow a division symbol?

parse

logical, default TRUE; should the units be made into an expression (so we get subscripts)? Setting to FALSE may be useful if parse fails, e.g. if the unit contains symbols that assume a particular encoding

Examples

units_options(sep = c("~~~", "~"), group = c("", "")) # more space, parenthesis
## set defaults:
units_options(sep = c("~", "~"), group = c("[", "]"), negative_power = FALSE, parse = TRUE)