Get information about valid units

The returned dataframe is constructed at runtime by reading the xml database that powers unit conversion in [package:udunits2]. Inspect this dataframe to determine what inputs are accepted by as_units (and the other functions it powers: make_units , set_units , units<-).

valid_udunits(quiet = FALSE)

valid_udunits_prefixes(quiet = FALSE)

Arguments

quiet
logical, defaults TRUE to give a message about the location of the udunits database being read.

Value

a data frame with columns symbol , symbol_aliases , name_singular , name_singular_aliases , name_plural , or name_plural_aliases , def , definition , comment , dimensionless and source_xml

Details

Any entry listed under symbol , symbol_aliases , name_singular , name_singular_aliases , name_plural , or name_plural_aliases is valid. Additionally, any entry under symbol or symbol_aliases may can also contain a valid prefix, as specified by valid_udunits_prefixes() .

Note, this is primarily intended for interactive use, the exact format of the returned dataframe may change in the future.

Examples

valid_udunits()

#> udunits system database read from /usr/share/xml/udunits

#> # A tibble: 275 x 11
#>    symbol symbol_aliases name_singular name_singular_aliases name_plural
#>     <chr>          <chr>         <chr>                 <chr>       <chr>
#>  1      m                        meter                 metre            
#>  2     kg                     kilogram                                  
#>  3      s                       second                                  
#>  4      A                       ampere                                  
#>  5      K                       kelvin                                  
#>  6    mol                         mole                                  
#>  7     cd                      candela                                  
#>  8    rad                       radian                                  
#>  9     sr                    steradian                                  
#> 10     Hz                        hertz                                  
#> # ... with 265 more rows, and 6 more variables: name_plural_aliases <chr>,
#> #   def <chr>, definition <chr>, comment <chr>, dimensionless <lgl>,
#> #   source_xml <chr>

valid_udunits_prefixes()

#> udunits system database read from /usr/share/xml/udunits

#> # A tibble: 20 x 4
#>    symbol symbol_aliases   name value
#>    <fctr>         <fctr> <fctr> <dbl>
#>  1      Y                 yotta 1e+24
#>  2      Z                 zetta 1e+21
#>  3      E                   exa 1e+18
#>  4      P                  peta 1e+15
#>  5      T                  tera 1e+12
#>  6      G                  giga 1e+09
#>  7      M                  mega 1e+06
#>  8      k                  kilo 1e+03
#>  9      h                 hecto 1e+02
#> 10     da                  deka 1e+01
#> 11      d                  deci 1e-01
#> 12      c                 centi 1e-02
#> 13      m                 milli 1e-03
#> 14      µ           μ, u  micro 1e-06
#> 15      n                  nano 1e-09
#> 16      p                  pico 1e-12
#> 17      f                 femto 1e-15
#> 18      a                  atto 1e-18
#> 19      z                 zepto 1e-21
#> 20      y                 yocto 1e-24

if(interactive())
  View(valid_udunits())

#> udunits system database read from /usr/share/xml/udunits