deparse_unit {units}	R Documentation

deparse unit to string in product power form (e.g. km m-2 s-1)

Description

deparse unit to string in product power form (e.g. km m-2 s-1)

Usage

deparse_unit(x)

as_cf(x)

Arguments

x
object of class units

Details

as_cf is deprecated; use deparse_unit.

Value

length one character vector

Examples

u = as_units("kg m-2 s-1", implicit_exponents = TRUE)
u
deparse_unit(u)

[Package units version 0.5-1 Index]