install_symbolic_unit {units}	R Documentation

Define new symbolic units

Description

Adding a symbolic unit allows it to be used in as_units, make_units and set_units. No installation is performed if the unit is already known by udunits.

Usage

install_symbolic_unit(chr, warn = TRUE)

uninstall_symbolic_unit(chr, all = FALSE)

Arguments

chr
a length 1 character vector that is the new unit name or symbol.

warn
warns if the supplied unit symbol is already a valid unit symbol recognized by udunits.

all
if TRUE, uninstalls all user defined custom symbolic units

[Package units version 0.5-1 Index]