Set measurement units on a numeric vector

Set measurement units on a numeric vector

Convert units

retrieve measurement units from units object

# S3 method for numeric
units(x) <- value

# S3 method for units
units(x) <- value

set_units(x, value, ...)

# S3 method for units
set_units(x, value, ...)

# S3 method for numeric
set_units(x, value, ...)

# S3 method for units
units(x)

Arguments

x
numeric vector, or object of class units

value
object of class units or symbolic_units, or in the case of set_units expression with symbols that can be resolved in ud_units (see examples).

...
ignored

Value

object of class units

the units method retrieves the units attribute, which is of class symbolic_units

Details

set_units is a pipe-friendly version of units<- that evaluates value first in the environment of ud_units.

Examples

x = 1:3
class(x)

#> [1] "integer"

units(x) <- with(ud_units, m/s) # valid

#> Error in eval(expr, envir, enclos): object 'm' not found

class(x)

#> [1] "integer"

y = 2:5
a <- with(ud_units, 1:3 * m/s)

#> Error in eval(expr, envir, enclos): object 'm' not found

units(a) <- with(ud_units, km/h)

#> Error in eval(expr, envir, enclos): object 'km' not found

a

#> Error in eval(expr, envir, enclos): object 'a' not found

# note that these units have NOT been defined or declared before:
set_units(1:5, N/m^2)

#> Error in eval(expr, envir, enclos): object 'N' not found

if (require(magrittr)) {
 1:5 %>% set_units(N/m^2)
 1:10 %>% set_units(m) %>% set_units(km)
}

#> Loading required package: magrittr

#> Error in eval(expr, envir, enclos): object 'N' not found