Set measurement units on a numeric vector

Set measurement units on a numeric vector

Convert units

retrieve measurement units from units object

# S3 method for numeric
units(x) <- value

# S3 method for units
units(x) <- value

# S3 method for logical
units(x) <- value

# S3 method for units
units(x)

Arguments

x
numeric vector, or object of class units

value
object of class units or symbolic_units, or in the case of set_units expression with symbols that can be resolved in ud_units (see examples).

Value

object of class units

the units method retrieves the units attribute, which is of class symbolic_units

Details

if value is of class units and has a value unequal to 1, this value is ignored unless units_options("simplifiy") is TRUE. If simplify is TRUE, x is multiplied by this value.

Examples

x = 1:3
class(x)

#> [1] "integer"

units(x) <- as_units("m/s")
class(x)

#> [1] "units"

y = 2:5
a <- set_units(1:3, m/s)
units(a) <- with(ud_units, km/h)

#> Error in eval(substitute(expr), data, enclos = parent.frame()): object 'km' not found

a

#> Units: [m*s^-1]
#> [1] 1 2 3

# convert to a mixed_units object:
units(a) = c("m/s", "km/h", "km/h")
a

#> Mixed units: km*h^-1 (2), m*s^-1 (1) 
#> 1 [m*s^-1], 7.2 [km*h^-1], 10.8 [km*h^-1]