valid_udunits {units}	R Documentation

Get information about valid units

Description

The returned dataframe is constructed at runtime by reading the xml database that powers unit conversion in [package:udunits2]. Inspect this dataframe to determine what inputs are accepted by as_units (and the other functions it powers: as_units , set_units , units<-).

Usage

valid_udunits(quiet = FALSE)

valid_udunits_prefixes(quiet = FALSE)

Arguments

quiet
logical, defaults TRUE to give a message about the location of the udunits database being read.

Details

Any entry listed under symbol , symbol_aliases , name_singular , name_singular_aliases , name_plural , or name_plural_aliases is valid. Additionally, any entry under symbol or symbol_aliases may can also contain a valid prefix, as specified by valid_udunits_prefixes() .

Note, this is primarily intended for interactive use, the exact format of the returned dataframe may change in the future.

Value

a data frame with columns symbol , symbol_aliases , name_singular , name_singular_aliases , name_plural , or name_plural_aliases , def , definition , comment , dimensionless and source_xml

Examples

if (requireNamespace("xml2", quietly = TRUE)) {
  valid_udunits()
  valid_udunits_prefixes()
  if(interactive())
    View(valid_udunits())
}

[Package units version 0.6-6 Index]